coverage

+++ easy-rules-core/src/test/java/org/jeasy/rules/api/FactsTest.java
@@ -39,7 +39,7 @@ public class FactsTest {
         facts.put("foo", 2);
 
         assertThat(facts).hasSize(1);
-        assertThat(facts.get("foo")).isEqualTo(2);
+        assertThat((int) facts.get("foo")).isEqualTo(2);
     }
 
     @Test
@@ -72,7 +72,7 @@ public class FactsTest {
     @Test
     public void get() throws Exception {
         facts.put("foo", 1);
-        assertThat(facts.get("foo")).isEqualTo(1);
+        assertThat((int) facts.get("foo")).isEqualTo(1);
     }
 
     @Test
